+++ src/contexts/index.ts
@@ -1,0 +1,3 @@
+// import {Initialize} from './navigatedScreens/Initialize';
+
+export {RootContext} from './RootContext';
