--- src/frames/ModalFrame.tsx
@@ -1,1 +1,0 @@
-export {};
