--- src/frames/index.ts
@@ -1,5 +1,0 @@
-// import {Initialize} from './navigatedScreens/Initialize';
-
-export {RootFrame} from './RootFrame';
-
-// export {Modal} from './Modal';
